Removed style object and classNames from Textfields.
These styles turned out not to be necessary so they were removed.
Also in the submitSearchInputs method I nested the call to /search inside
of the response from /location. This is so that I can transform the input
from the user into a lat-long floating point number, once I have the
number then we can continue the submission to the /search endpoint.