Moved the router and component selector to the App.js component.
The Main component is a functional component that has one state hook
which is capturing the first call to action on the landing page. The
input element is just above the fold and will handle a quick search
for the closest doctors by just entering a zip code.
We should prompt the user that more advanced searching/filtering is
available when they get the results on the Body component.: