Reformatted the data for the autofilter in the Search Component.
Now each of the three files will follow the format of having one
variable that is a plain array which contains string that are sorted
alphabetically. This will be what is displayed in the auto filter
element to the right of the input boxes. And the second variable will
have additional details about the string in the first variable. Most
notably is the 'uid' string which can be used to request very specific
results from the Better Doctor API.