Adding a method to remove script tags.
When I retrieve a google map, they inject script tags to handle the
logic for the map. But when we search again the old script tags are
still in the head tag and they become doubled and tripled. So we are
adding logic to remove the tags when the component is unmounted.